Rewrite sv_envsense in C++ style, store data in BLE format immediately

+++ arduino/include/sensors/data.h
@@ -3,19 +3,21 @@
 
 #include <Arduino.h>
 
+#include <common.h>
+
 namespace sensors {
   struct sensor_data {
     // normalized BME688 sensor values provided by the `bsec2` library
-    float air_pressure;
-    float air_quality;
-    float humidity;
-    float temperature;
+    uint32_t air_pressure;
+    uint16_t air_quality;
+    uint16_t humidity;
+    int16_t temperature;
 
     // normalized illuminance (0..n lx) read from light sensor (see `light.cpp`)
-    unsigned int illuminance;
+    uint24 illuminance;
 
     // normalized soil moisture (0..100%) read from light sensor (see `light.cpp`)
-    unsigned int soil_moisture;
+    uint8_t soil_moisture;
   };
 
   extern struct sensor_data current_data;
